requirements.txt
python-telegram-bot==13.7
playwright==1.17.1
